+++ fresh.gen.ts
@@ -2,7 +2,7 @@
 // This file SHOULD be checked into source version control.
 // This file is automatically updated during development when running `dev.ts`.
 
-import * as $_userid_index from "./routes/[userid]/index.tsx";
+import * as $_userId_index from "./routes/[userId]/index.tsx";
 import * as $_app from "./routes/_app.tsx";
 import * as $index from "./routes/index.tsx";
 
@@ -10,7 +10,7 @@ import { type Manifest } from "$fresh/server.ts";
 
 const manifest = {
   routes: {
-    "./routes/[userid]/index.tsx": $_userid_index,
+    "./routes/[userId]/index.tsx": $_userId_index,
     "./routes/_app.tsx": $_app,
     "./routes/index.tsx": $index,
   },
